refactor: 调整 dtos 的目录结构
