--- src/auth/dtos/register-request.dto.ts
@@ -1,1 +1,0 @@
-
